Update build.yml and create-release job configurations
Updated the `jobs` section in `build.yml` to specify `osx-installer` in the `with` clause. Added a new condition to the `create-release` job to run only if the GitHub reference is `refs/heads/main`. The `create-release` job now depends on `build-windows` and `build-osx` jobs and runs on the latest Ubuntu version. A new step `Checkout code` has been added to the `create-release` job to checkout the code using `actions/checkout@v4`.